feat(login flow): change the login flow of the application
User will first see the main page where the user can navigate to
login page, activities, or forums
