+++ public/index.html
@@ -5,7 +5,7 @@
     <link rel="icon" href="%PUBLIC_URL%/app_logo.png" />
     <meta name="viewport" content="width=device-width, initial-scale=1" />
     <meta name="theme-color" content="#000000" />
-    <meta name="Advanced Analytical Tool" content="Visualize your data" />
+    <meta name="Asturec" content="Improving your child needs" />
     <link rel="apple-touch-icon" href="%PUBLIC_URL%/react.png" />
     <!--
       manifest.json provides metadata used when your web app is installed on a
@@ -21,7 +21,7 @@
       work correctly both with client-side routing and a non-root public URL.
       Learn how to configure a non-root public URL by running `npm run build`.
     -->
-    <title>Advanced Analytical Tool</title>
+    <title>Asturec</title>
     <script
       defer
       src="https://use.fontawesome.com/releases/v5.3.1/js/all.js"
