+++ src/app/App.tsx
@@ -1,9 +1,20 @@
 import React from 'react';
 import AppLoader from 'components/appLoader';
 import { useUser } from 'contexts/UserContext';
+import { BrowserRouter as Router, Switch, Route } from 'react-router-dom';
+import AppLayout from 'layouts/AppLayout';
+import Main from 'routes/main';
+import { HOME } from 'constants/routes';
 
 const UnauthenticatedApp = React.lazy(() => import('./UnauthenticatedApp'));
 const AuthenticatedApp = React.lazy(() => import('./AuthenticatedApp'));
+
+const ScrollToTop = () => {
+  React.useEffect(() => {
+    window.scrollTo(0, 0);
+  }, []);
+  return null;
+};
 
 const App: React.FC = () => {
   const { user } = useUser();
@@ -13,7 +24,15 @@ const App: React.FC = () => {
   }
   return (
     <React.Suspense fallback={<AppLoader />}>
-      {uid ? <AuthenticatedApp /> : <UnauthenticatedApp />}
+      <Router>
+        <AppLayout>
+          <ScrollToTop />
+          <Switch>
+            <Route exact path={HOME} component={Main} />
+            {uid ? <AuthenticatedApp /> : <UnauthenticatedApp />}
+          </Switch>
+        </AppLayout>
+      </Router>
     </React.Suspense>
   );
 };
