+++ src/app/UnauthenticatedApp.tsx
@@ -3,33 +3,28 @@ import {
   BrowserRouter as Router,
   Switch,
   Route,
-  useLocation
+  useHistory
 } from 'react-router-dom';
 
-import Main from 'routes/main';
-import AppLayout from 'layouts/AppLayout';
+import Login from 'routes/login';
 
-import { HOME } from 'constants/routes';
+import { LOGIN, HOME } from 'constants/routes';
 
-const ScrollToTop = () => {
-  const { pathname } = useLocation();
-
+const RedirectPath = () => {
+  const history = useHistory();
   React.useEffect(() => {
-    window.scrollTo(0, 0);
-  }, [pathname]);
-
+    history.push(HOME);
+  });
   return null;
 };
 
 const UnauthenticatedApp: React.FC = () => {
   return (
     <Router>
-      <AppLayout>
-        <ScrollToTop />
-        <Switch>
-          <Route exact path={HOME} component={Main} />
-        </Switch>
-      </AppLayout>
+      <Switch>
+        <Route exact path={LOGIN} component={Login} />
+        <Route component={RedirectPath} />
+      </Switch>
     </Router>
   );
 };
