+++ src/constants/routes.ts
@@ -1,1 +1,3 @@
 export const HOME = '/';
+export const LOGIN = '/login';
+export const PROFILE = '/profile';
