+++ src/routes/login/Login.scss
@@ -1,0 +1,35 @@
+@import 'assets/scss/theme.scss';
+
+.login {
+  &__box {
+    -webkit-box-shadow: 0px 0px 5px 1px rgba(0, 0, 0, 0.75);
+    -moz-box-shadow: 0px 0px 5px 1px rgba(0, 0, 0, 0.75);
+    box-shadow: 0px 0px 5px 1px rgba(0, 0, 0, 0.75);
+    border-radius: 20px;
+    padding-top: 20px;
+    padding-bottom: 20px;
+    margin-top: 50px;
+    min-height: 500px;
+
+    &--icon {
+      display: block;
+      margin: auto;
+    }
+
+    &--org-name {
+      text-align: center;
+      font-size: 30px;
+    }
+
+    &--wrapper {
+      margin-top: 50px;
+      margin-left: 10px;
+      margin-right: 10px;
+      margin-bottom: 50px;
+    }
+  }
+
+  &__content {
+    margin-top: 100px;
+  }
+}
